Dev phase modular backend v.0.8
