+++ backend/python_scripts/RestfulAPI.py
@@ -1,8 +1,13 @@
+# Pre-defined Libraries
 from flask import Flask, request, jsonify
 import pandas as pd
-import sqlite3
+import datetime
 import requests
 import os
+
+# Defined Libraries
+import paths_logging
+import DBoperations
 
 app = Flask(__name__)
 
@@ -11,13 +16,12 @@ app = Flask(__name__)
 def index():
     return "Hello, I am flask API, up  and running."
 
-# Get the current directory of the Flask app
-current_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-
-# Define a relative path to the Excel file
-file_path = os.path.join(current_dir, 'saved data', 'Data for 24-Aug-2023', 'Nifty_Data_24-Aug-2023.xlsx')
 
 # Read Excel data and convert to JSON
+folderPath = paths_logging.savedData_folder_path()
+# Define a relative path to the Excel file
+file_path = os.path.join(folderPath, f"Data for {datetime.date.today().strftime('%d-%b-%Y')}", 'Nifty_Data_31-Aug-2023.xlsx')
+
 df = pd.read_excel(file_path)
 df_json = df.to_json(orient="split")
 
@@ -40,56 +44,14 @@ def get_chart_data():
     return jsonify(chart_data)
 
 
-folder_path = os.path.join(current_dir,"DB files")
-
-if not os.path.exists(folder_path):
-    try:
-        os.makedirs(folder_path)
-    except Exception as e:
-        print(e)
-
-DB_file_path = os.path.join(folder_path,'strike_prices.db')
 # Initialize the SQLite database
-conn = sqlite3.connect(DB_file_path)
-cursor = conn.cursor()
-# Create a table to store nf_SP and bnf_SP
-cursor.execute('''
-    CREATE TABLE IF NOT EXISTS strike_prices (
-        id INTEGER PRIMARY KEY,
-        dateTime DATETIME,
-        nf_SP REAL,
-        bnf_SP REAL
-    )
-''')
-conn.commit()
-conn.close()
-
-def fetch_strike_prices():
-    try:
-        conn = sqlite3.connect(DB_file_path)
-        cursor = conn.cursor()
-
-        # Retrieve the nf_SP and bnf_SP values from the database
-        cursor.execute('SELECT dateTime, nf_SP, bnf_SP FROM strike_prices ORDER BY dateTime DESC LIMIT 1')
-        result = cursor.fetchone()
-
-        conn.close()
-
-        if result:
-            dateTime, nf_SP, bnf_SP = result
-            return dateTime, nf_SP, bnf_SP
-        else:
-            return None, None, None
-
-    except Exception as e:
-        raise e  # You can log or handle the exception as needed
-
+DBoperations.createTableIfNotExists()
 
 @app.route('/api/strikePrice')
 def get_strike_price():
     try:
         # Retrieve nf_SP and bnf_SP from the database
-        dateTime, nf_SP, bnf_SP = fetch_strike_prices()
+        dateTime, nf_SP, bnf_SP = DBoperations.fetch_strike_prices()
 
         if nf_SP is not None and bnf_SP is not None:
             return jsonify({'dateTime': dateTime,'nf_SP': nf_SP, 'bnf_SP': bnf_SP})
